+++ src/sstc_core/sites/spectral/config/db_schemas/platforms_schemas.py
@@ -1,0 +1,7 @@
+PhenoCams = [
+    
+    {
+    'field_name': 'L2_RGB_CI_SUM_filepath',
+    'field_type': 'VARCHAR',
+    'field_default_vale': None},
+]