+++ src/sstc_core/sites/spectral/stations/abisko.py
@@ -17,11 +17,12 @@ meta = {
     "platforms_filepath": config_dirpath / 'platforms' / 'platforms_abisko.yaml',
     "phenocam_quality_weights_filepath": config_dirpath / "phenocam_quality_weights.yaml",    
     'geolocation':{
-        'point':{
-            'epsg:4326': {'latitude_dd': None, 'longitude_dd': None}, 
-            'epsg:3006': {'x_coord': None, 'y_coord': None}},
+        'point':{ 
+          'epsg': "epsg:4326",
+          'latitude_dd': 68.353729,
+          'longitude_dd': 18.816522,
     }
-    
+    } 
     }
 
 def load_configurations():